upadte new

+++ database/migrations/0001_01_01_000002_create_jobs_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('queue');
             $table->longText('payload');
-            $table->unsignedTinyInteger('attempts');
+            $table->smallInteger('attempts');
             $table->unsignedInteger('reserved_at')->nullable();
             $table->unsignedInteger('available_at');
             $table->unsignedInteger('created_at');
